REF: move separable stuff out of atef.checks
